Use @goto-bus-stop/common-shake
This includes a PR to track dependent uses. That means that, if an
export is used exclusively in unused functions, it will also get
removed.

This is especially helpful for things like `util` and `path`, which have
a lot of interdependent functions that could previously not be removed.
For example, `util.format` was always included because it is used in
other functions in the `util` module, even if the only thing the bundle
used was `util.inherits`.